% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{aka}
\alias{aka}
\title{"Also known as" splits collapsed list of gene identifiers and tests if
the expression is part of it (could be replaced by `str_detect()` or similar)}
\usage{
aka(x, y, sep = " */+ *")
}
\arguments{
\item{x}{Collapsed list of genes (string)}

\item{y}{Value to search for}

\item{sep}{the separator to be used (defaults to ` */+ *`)}
}
\description{
"Also known as" splits collapsed list of gene identifiers and tests if
the expression is part of it (could be replaced by `str_detect()` or similar)
}
