% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feature_data}
\alias{feature_data}
\title{Extract available features as a tibble
A wrapper around rowData() for SummarizedExperiment
and fData() for ExpressionSet objects
returning a tibble}
\usage{
feature_data(x, ...)
}
\arguments{
\item{x}{A SummarizedExperiment or ExpressionSet object.}

\item{...}{Arguments to be passed to methods}
}
\description{
Extract available features as a tibble
A wrapper around rowData() for SummarizedExperiment
and fData() for ExpressionSet objects
returning a tibble
}
