% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{filter_features}
\alias{filter_features}
\title{Filters out expression data from
SummarizedExperiment or ExpressionSet objects
matching feature criteria}
\usage{
filter_features(x, ...)
}
\arguments{
\item{x}{A SummarizedExperiment or ExpressionSet object.}

\item{...}{Arguments passed to filter}
}
\description{
Filters out expression data from
SummarizedExperiment or ExpressionSet objects
matching feature criteria
}
