% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probesets.R
\name{filter_probeset}
\alias{filter_probeset}
\title{Identify probeset exhibiting the highest expression variance
and being the most specific (based on probeset IDs)}
\usage{
filter_probeset(x, feature_var, most_specific = TRUE, ...)
}
\arguments{
\item{x}{An ExpressionSet or SummarizedExperiment object}

\item{feature_var}{The gene identifier to group by and associated to
multiple probesets (row IDs)}

\item{most_specific}{remove suboptimal affymetrix probesets (`at > a_at > s_at > x_at`). Defaults to `TRUE`.}

\item{...}{Arguments to be passed to methods}
}
\description{
Identify probeset exhibiting the highest expression variance
and being the most specific (based on probeset IDs)
}
