% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{filter_samples}
\alias{filter_samples}
\title{Filter out samples of interest
based on col data variables (or .sample as the column identifier)}
\usage{
filter_samples(x, ...)
}
\arguments{
\item{x}{A SummarizedExperiment or ExpressionSet object.}

\item{...}{Arguments passed to filter}
}
\description{
Filter out samples of interest
based on col data variables (or .sample as the column identifier)
}
