% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarized_experiment.R
\name{get_goi}
\alias{get_goi}
\title{Function to extract expression data from SummarizedExperiment objects}
\usage{
get_goi(x, goi, ...)
}
\arguments{
\item{x}{A SummarizedExperiment object.}

\item{goi}{A character vector containing the gene symbols of interest}
}
\description{
Function to extract expression data from SummarizedExperiment objects
}
