% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{\%aka\%}
\alias{\%aka\%}
\title{"Also known as" splits collapsed list of gene identifiers and tests if
the expression is part of it (could be replaced by `str_detect()` or similar)}
\usage{
x \%aka\% y
}
\arguments{
\item{x}{Collapsed list of genes (string)}

\item{y}{Value to search for}
}
\description{
"Also known as" splits collapsed list of gene identifiers and tests if
the expression is part of it (could be replaced by `str_detect()` or similar)
}
