% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_transform.R
\name{log_transform}
\alias{log_transform}
\title{Log transform a vector adding and add a constant}
\usage{
log_transform(x, i = min(x[x > 0])/2, base = 2)
}
\arguments{
\item{x}{A numeric vector.}

\item{i}{Numeric value to add before performing the log-transformation. By default half of the smallest non-zero value is added.}

\item{base}{A positive or complex number: the base with respect to which logarithms are computed. Defaults to 2.}
}
\description{
Log transform a vector adding and add a constant
}
