% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probesets.R
\name{select_probeset}
\alias{select_probeset}
\title{Identify probeset exhibiting the highest expression variance
and being the most specific (based on probeset IDs)}
\usage{
select_probeset(x, gene_id_var = gene_symbol, ...)
}
\arguments{
\item{x}{An ExpressionSet or SummarizedExperiment object}

\item{gene_id_var}{The gene identifier to group by and associated to
multiple probesets (row IDs)}
}
\description{
Identify probeset exhibiting the highest expression variance
and being the most specific (based on probeset IDs)
}
