% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{semi_join_samples}
\alias{semi_join_samples}
\title{Filter out samples of interest using semi_join()}
\usage{
semi_join_samples(x, y, by = NULL, ...)
}
\arguments{
\item{x}{A SummarizedExperiment or ExpressionSet object.}

\item{y}{A tibble or data.frame}

\item{by}{A character vector of variables to join by.}

\item{...}{Arguments passed to filter}
}
\description{
Filter out samples of interest using semi_join()
}
