% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{filter_features}
\alias{filter_features}
\title{Filters out expression data from
SummarizedExperiment or ExpressionSet objects
matching feature criteria}
\usage{
filter_features(x, e, ...)
}
\arguments{
\item{x}{A SummarizedExperiment or ExpressionSet object.}

\item{e}{An expression to filter features based on available variables}

\item{...}{Arguments to be passed to methods}
}
\description{
Filters out expression data from
SummarizedExperiment or ExpressionSet objects
matching feature criteria
}
