% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{filter_samples}
\alias{filter_samples}
\title{Filter out samples of interest
based on col data variables (or .sample as the column identifier)}
\usage{
filter_samples(x, e, ...)
}
\arguments{
\item{x}{A SummarizedExperiment or ExpressionSet object.}

\item{e}{An expression to filter features based on available variables}

\item{...}{Arguments to be passed to methods}
}
\description{
Filter out samples of interest
based on col data variables (or .sample as the column identifier)
}
